Add sandbox mode
Adding sandbox just enableing a internal variable _sandbox which is set by the config object passed by the constructor. 
This Variable is also passed when signing the request (so the url matches the requested one)
Also catching error when sandbox mode is on and error.code is 'InternalFailure' this means sandbox parameters are not correct.